QPID-1951: Removed need for Windows versions of ssize_t and pid_t - Trivially removed Windows uses of ssize_t - Rearchitected how the Windows port finds an existing qpidd to stop it - Split Posix Lockfile functionality using pids into a new PidFile class
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@890929 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/tests/SocketProxy.h
@@ -141,12 +141,12 @@ class SocketProxy : private qpid::sys::Runnable
                 }
                 // Something is set; relay data as needed until something closes
                 if (FD_ISSET(server->getFd(), &socks)) {
-                    ssize_t n = server->read(buffer, sizeof(buffer));
+                    int n = server->read(buffer, sizeof(buffer));
                     throwIf(n <= 0, "SocketProxy: server disconnected");
                     if (!dropServer) client.write(buffer, n);
                 }
                 if (FD_ISSET(client.getFd(), &socks)) {
-                    ssize_t n = client.read(buffer, sizeof(buffer));
+                    int n = client.read(buffer, sizeof(buffer));
                     throwIf(n <= 0, "SocketProxy: client disconnected");
                     if (!dropServer) server->write(buffer, n);
                 }
